Fixed gcc warnings
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@4618 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ src/media_tools/mpd.c
@@ -181,7 +181,7 @@ static u64 gf_mpd_parse_date(char *attr)
 		res = res + diff;
 	}
 	res *= 1000;
-	ms = (s - (u32)s) * 1000;
+	ms = (u32) ( (s - (u32)s) * 1000);
 	return res + ms;
 }
